+++ version/112/frontend/heidelpay.hook140.php
@@ -39,5 +39,44 @@ if (preg_match('/[0-9]{3}\.[0-9]{3}\.[0-9]{3}/', $_GET ['hperror'])) {
     $errorSnip = $divStart . ' ' . $hpErrorMsg . ' ' . $divEnd;
     pq("#content")->prepend($errorSnip);
 
-    unset($_SESSION['heidelpayErrorCode']);
+}
+
+if(isset($_GET ['hperrorcom'])){
+
+    $divStart = '<div class="alert alert-danger"><strong>Error:</strong><br>';
+    $divEnd = '</div>';
+
+
+    if($_SESSION ['cISOSprache'] == 'ger')
+    {
+        $hpErrorMsg = utf8_decode('Dieses Zahlverfahren steht nicht für Firmenkunden zur Verfügung');
+    }else{
+        $hpErrorMsg = utf8_decode('This paymentmethod is not available for corporate clients');
+    }
+
+
+
+    $errorSnip = $divStart . ' ' . $hpErrorMsg . ' ' . $divEnd;
+    pq("#content")->prepend($errorSnip);
+
+}
+
+if(isset($_GET ['hperroradd'])){
+
+    $divStart = '<div class="alert alert-danger"><strong>Error:</strong><br>';
+    $divEnd = '</div>';
+
+
+    if($_SESSION ['cISOSprache'] == 'ger')
+    {
+        $hpErrorMsg = utf8_decode('Rechnungs- und Lieferadresse müssen identisch sein');
+    }else{
+        $hpErrorMsg = utf8_decode('Billing- and shipping adress have to be equal');
+    }
+
+
+
+    $errorSnip = $divStart . ' ' . $hpErrorMsg . ' ' . $divEnd;
+    pq("#content")->prepend($errorSnip);
+
 }