+++ CHANGELOG.md
@@ -1,12 +1,28 @@
 # Release Notes - heidelpay extension for JTL Shop
 
 ## v112
+17.05.03:
+
+JTL plugin version 112 introduces payment via white label gateway. 
+All former payment methods were moved to white label and new payment methods were added.
 
 ### Added
+- version folder 112
+- new payment method Przelewy24
+- new payment method Postfinance eFinance
+- new payment method Postfinance Card
+- new payment method Santander invoice
+- new payment method invoice secured
+- new payment method direct debit secured
 
 ### Changed
-
+- updated version to 112
+- existing payment methods now running via white label gateway 
 
 ### Fixed
+- notice 'not supported for JTL4' doesn't appear anymore 
 
 ### Removed
+- sql folder
+- payment method Billsafe
+- unused code